requirements.txt
google-generativeai
pandas
numpy
scikit-learn
streamlit
python-dotenv
requests
beautifulsoup4